weather_service/requirements.txt
fastapi
requests
fastapi-cache2